refactor: add provider-neutral player coordinator
