+++ resources/default-settings/jukebox.default.yaml
@@ -11,7 +11,7 @@ modules:
     jingle: jingle
     jingle.alsawave: jingle.alsawave
     jingle.jinglemp3: jingle.jinglemp3
-    player: playermpd
+    player: player.plugin
     cards: rfid.cards
     rfid: rfid.reader
     timers: timers
