+++ src/jukebox/components/player/backends/__init__.py
@@ -1,0 +1,1 @@
+"""Playback backend implementations used by the player coordinator."""
