+++ src/jukebox/components/synchronisation/rfidcards/__init__.py
@@ -19,7 +19,6 @@ The local copy is kept to reduce unnecessary syncing.
 import logging
 import subprocess
 import components.player
-import components.playermpd
 import components.rfid.reader
 import components.synchronisation.syncutils as syncutils
 import jukebox.cfghandler
@@ -29,7 +28,7 @@ import os
 import shutil
 
 from components.rfid.reader import RfidCardDetectState
-from components.playermpd.playcontentcallback import PlayCardState
+from components.player.playcontentcallback import PlayCardState
 
 
 logger = logging.getLogger('jb.sync_rfidcards')
@@ -73,7 +72,7 @@ class SyncRfidcards:
                     self._sync_remote_ssh_user = cfg_sync_rfidcards.getn('sync_rfidcards', 'credentials', 'username')
 
             components.rfid.reader.rfid_card_detect_callbacks.register(self._rfid_callback)
-            components.playermpd.play_card_callbacks.register(self._play_card_callback)
+            components.player.play_card_callbacks.register(self._play_card_callback)
         else:
             logger.info("Sync RFID cards deactivated")
